# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  contracts.DirectoryDownloadRequest:
    properties:
      auto_classify:
        type: boolean
      directory_path:
        type: string
      recursive:
        type: boolean
      target_dir:
        type: string
      video_only:
        type: boolean
    required:
    - directory_path
    type: object
...
  contracts.FileClassificationRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/contracts.FileResponse'
        type: array
    required:
    - files
    type: object
  contracts.FileDownloadRequest:
    properties:
      auto_classify:
        type: boolean
      file_path:
        type: string
      options:
        additionalProperties: true
        type: object
      target_dir:
        type: string
    required:
    - file_path
    type: object
  contracts.FileListRequest:
    properties:
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 1000
        minimum: 1
        type: integer
      path:
        type: string
      recursive:
        type: boolean
      sort_by:
        enum:
        - name
        - size
        - modified
        type: string
      sort_order:
        enum:
        - asc
        - desc
        type: string
      video_only:
        type: boolean
    required:
    - path
    type: object
...
  contracts.FileResponse:
    properties:
      category:
        type: string
      download_path:
        type: string
      external_url:
        type: string
      internal_url:
        type: string
      is_dir:
        type: boolean
      media_type:
        type: string
      modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      size_formatted:
        type: string
      thumbnail:
        type: string
    type: object
  contracts.FileSearchRequest:
    properties:
      file_type:
        type: string
      limit:
        maximum: 1000
        minimum: 1
        type: integer
      max_size:
        type: integer
      min_size:
        type: integer
      modified_after:
        type: string
      modified_before:
        type: string
      path:
        type: string
      query:
        type: string
    required:
    - query
    type: object
...
  contracts.TimeRangeFileRequest:
    properties:
      end_time:
        type: string
      hours_ago:
        maximum: 8760
        minimum: 1
        type: integer
      path:
        type: string
      start_time:
        type: string
      video_only:
        type: boolean
    required:
    - end_time
    - path
    - start_time
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin框架的Alist和Aria2下载管理服务
  license:
    name: MIT
    url: https://github.com/easayliu/alist-aria2-download/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Alist Aria2 Download API
  version: "1.0"
paths:
...
  /files/category/{category}:
    get:
      description: 获取指定分类的文件列表
      parameters:
      - description: 文件分类
        enum:
        - video
        - audio
        - image
        - document
        - archive
        - other
        in: path
        name: category
        required: true
        type: string
      - description: 搜索路径
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分类文件列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 按分类获取文件
      tags:
      - 文件管理
  /files/classify:
    post:
      consumes:
      - application/json
      description: 将文件列表按照类型进行分类
      parameters:
      - description: 文件分类请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.FileClassificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分类结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 分类文件
      tags:
      - 文件管理
  /files/download:
    post:
      consumes:
      - application/json
      description: 获取指定路径下的所有文件并添加到Aria2下载队列，支持递归下载子目录
      parameters:
      - description: 下载路径请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.DirectoryDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 下载任务创建结果或预览信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 从指定路径下载文件
      tags:
      - 文件管理
  /files/list:
    post:
      consumes:
      - application/json
      description: 获取指定路径下的文件列表，支持分页和视频文件过滤
      parameters:
      - description: 列出文件请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.FileListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 列出指定路径的文件
      tags:
      - 文件管理
  /files/manual-download:
    post:
      consumes:
      - application/json
      description: 手动执行指定时间范围内的文件下载，支持预览模式
      parameters:
      - description: 下载参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TimeRangeFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 下载结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 手动执行文件下载
      tags:
      - 文件管理
  /files/recent:
    get:
      consumes:
      - application/json
      description: 获取最近指定小时内修改的文件
      parameters:
      - description: 搜索路径
        in: query
        name: path
        type: string
      - default: 24
        description: 小时数
        in: query
        name: hours_ago
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 最近文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取最近文件
      tags:
      - 文件管理
  /files/search:
    post:
      consumes:
      - application/json
      description: 在指定路径中搜索符合条件的文件
      parameters:
      - description: 搜索请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.FileSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 搜索文件
      tags:
      - 文件管理
  /files/single-download:
    post:
      consumes:
      - application/json
      description: 下载指定路径的单个文件到Aria2
      parameters:
      - description: 文件下载请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.FileDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 文件下载任务创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 下载单个文件
      tags:
      - 文件管理
  /files/time-range:
    post:
      consumes:
      - application/json
      description: 获取指定时间范围内修改的文件列表
      parameters:
      - description: 时间范围请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TimeRangeFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 时间范围内的文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 按时间范围获取文件
      tags:
      - 文件管理
  /files/yesterday:
    get:
      consumes:
      - application/json
      description: 获取昨天修改的文件列表
      parameters:
      - description: 搜索路径（留空使用配置的默认路径）
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 昨天的文件列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取昨天的文件
      tags:
      - 文件管理
  /files/yesterday/download:
    post:
      consumes:
      - application/json
      description: 将昨天修改的文件批量添加到Aria2下载队列
      parameters:
      - description: 搜索路径（留空使用配置的默认路径）
        in: query
        name: path
        type: string
      - description: 预览模式，只生成路径不下载
        in: query
        name: preview
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 下载任务创建结果或预览信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量下载昨天文件
      tags:
      - 文件管理
...
schemes:
- http
- https
swagger: "2.0"